Bug#37460061 - CLI: PREVENTS ENTRY OF CORRECT ROOT PASSWORD, WHEN IT'S SHORT
Description: A similar bug was raised for Installer (29036252) and was duplicated for Configurator. The original issue was that in Installer the password of the root user was not allowed to be short (lower than 4 characters) when performing a connection validation during a Router configuration. For Installer it was agreed that we allow empty and short passwords when performing connection attempts (similar behavior exists in the MySQL client) but continue to have the existing validation rules to require a password and said password be at least 4 characters long during a new configuration. As part of this fix for Configurator, it was found that the GUI already works as expected but the CLI did not, that is why the bug name was updated to only mention CLI. Additionally, automated tests were added to validate consistent behavior between GUI and CLI in this regard.

- Added a new exit code (RootPasswordInvalidFormat) used to be shown whenever validation of the root password fails, it is prepared to include the message generated by the validation (Configurator/Base/Enums/ExitCode.cs).
- Updated the ProcessPasswordOption method to perform the same password validations as the GUI for the root password whenever a new configuration is taking place. Also updated the same method to only require the Password CLI option during a new configuration since not providing it in other configuration modes indicates that the password is empty. This aligns with the CLI behavior of the GUI and the MySQL client in which empty passwords are allowed (Configurator/Core/CLI/CommandLine.cs).
- Updated the validator tool to support sending empty passwords when validating a connection (tests/bin/mysql_validator.exe, tests/tools/mysql_validator/Program.cs).
- Updated the configureAccountsAndRolesPage function to support not clicking the Next button on the Accounts and Roles configuration page. This is needed in the new "Minimum length for passwords validation" GUI test which validates that the Next button is disabled due to an invalid password after having configured the Accounts and Roles page (tests/lib/mga-functions.js).
- Added additional logging for whenever an SQL command is executed (tests/lib/test-runner.js).
- Added GUI/CLI tests to validate consistent behavior of not allowing empty or passwords shorter than 4 characters during a new configuration but allowing empty or short password during connection attempts since the user may have changed the password outside of MySQL Configurator (tests/tests/cli.js, tests/tests/regression.js).

Change-Id: I4e00f06eac60afccf820a5d07d2ee89c74f8f506